Break up step definition file
